requirements.txt
asgiref==3.5.0
astroid==2.11.3
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.4
Django==4.0.3
django-filter==21.1
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.11.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
openapi-codec==1.3.2
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.13.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
yapf==0.32.0
